\name{data_Rep2}
\docType{data}
\alias{data_Rep2}
\title{example of data}
\description{
A short example dataset for ChIPmix or multiChIPmix analysis.
}

\format{
It is a text file containing 98 ordered rows and 3 columns. The first column corresponds to the identifier, the second column is the INPUT intensity value and the third column is the IP intensity value.
}

\keyword{datasets}
