\name{dataHMM_Rep1}
\docType{data}
\alias{dataHMM_Rep1}
\title{example of data}
\description{
A short example dataset for ChIPmix or multiChIPmix analysis in the case of dependency.
}

\format{
It is a text file containing 98 ordered rows and 3 columns. The first column corresponds to the identifier, the second column is the INPUT intensity value and the third column is the IP intensity value.
}

\keyword{datasets}
